api/requirements.txt
Django==1.10.3
djangorestframework==3.5.3
mysqlclient==1.3.9
olefile==0.43
Pillow==4.0.0
